requirements.txt
av==15.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
cython==3.1.2
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jinja2==3.1.6
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyvista==0.45.2
pyyaml==6.0.2
requests==2.32.4
rtree==1.4.0
scooby==0.10.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
trimesh==4.6.13
triton==3.3.1
typing-extensions==4.14.1
urllib3==2.5.0
vtk==9.4.2
